- вынесла классы в отдельные файлы - поправила ф-ию вычисления окончания - проверила, что не вводят неккоректные значения
