+++ src/main/java/Product.java
@@ -1,0 +1,9 @@
+class Product {
+    String name;
+    double price;
+
+    Product(String name, double price) {
+        this.name = name;
+        this.price = price;
+    }
+}
